display the current counts of the single channel selected to display
git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@278 71953026-4e7b-de11-b703-00e01807f32f

+++ lib/histogram.h
@@ -56,7 +56,7 @@ public:
 	quint16 maxpos() {return m_maximumPos;}
 
 	//! \return sum of counts in the spectrum
-	quint16 getTotalCounts() {return m_totalCounts;}
+	quint64 getTotalCounts() {return m_totalCounts;}
 
 	float mean(float &s);
